Removing redundant xhr event listeners.
During development, it was noted that <div>s corresponding to courses
would be deleted and then re-constructed (with new IDs) whenever the
cart reloaded and whenever the window resized. The first event was
originally identified via a pseudo-event-listener that tied into the
onreadystatechange event through the xmlHttpRequest prototype; the
second event was tied into with a MutationObserver(), which as it turned
out, is sufficient to describe *both* events. When dev came back to
restructure the xhr handling - to minimize security risks and to
simplify command structure - it was found that xhr handling was simply
unnecessary and moreover actually rather inefficient. As such, it's now
being deleted.